Fixed decimal precision to 10 places

+++ ddl.sql
@@ -7,10 +7,10 @@ DROP TABLE IF EXISTS probability;
 
 CREATE TABLE IF NOT EXISTS probability
 (
-    id                                              INT     AUTO_INCREMENT PRIMARY KEY,
-    probability_of_initial_knowledge                DECIMAL NOT NULL,
-    individualized_probability_of_initial_knowledge DECIMAL NOT NULL,
-    probability_of_learning                         DECIMAL NOT NULL,
-    probability_of_guess                            DECIMAL NOT NULL,
-    probability_of_slip                             DECIMAL NOT NULL
+    id                                              INT             AUTO_INCREMENT PRIMARY KEY,
+    probability_of_initial_knowledge                DECIMAL(11, 10) NOT NULL,
+    individualized_probability_of_initial_knowledge DECIMAL(11, 10) NOT NULL,
+    probability_of_learning                         DECIMAL(11, 10) NOT NULL,
+    probability_of_guess                            DECIMAL(11, 10) NOT NULL,
+    probability_of_slip                             DECIMAL(11, 10) NOT NULL
 );
